Added ability to disable elastic service

+++ Backend/services/elastic.js
@@ -1,12 +1,16 @@
 let elasticsearch = require('elasticsearch');
 
-var { connectionString, indexPrefix } = require('../config/environment.js').elastic;
+var { enable, connectionString, indexPrefix } = require('../config/environment.js').elastic;
 
-let client = new elasticsearch.Client({
+let client;
+
+if (enable) client = new elasticsearch.Client({
   hosts: [connectionString]
 });
 
 let index = (index, document) => {
+  if (!enable) return Promise.resolve();
+
   if (document.toJSON) document = document.toJSON();
 
   return client.index({
@@ -18,6 +22,8 @@ let index = (index, document) => {
 };
 
 let remove = (index, docId) => {
+  if (!enable) return Promise.resolve();
+
   return client.delete({
     index: indexPrefix + index,
     type: index,
@@ -26,6 +32,8 @@ let remove = (index, docId) => {
 };
 
 let bulk = (action, index, documents) => {
+  if (!enable) return Promise.resolve();
+
   let query = documents.reduce((acc, document) => {
     acc.push({
       [action]: {
@@ -49,6 +57,8 @@ let bulk = (action, index, documents) => {
 };
 
 let search = (index, userId, queryString) => {
+  if (!enable) throw new Error("ElasticSearch not enabled");
+
   return client.search({
     index: indexPrefix + index,
     type: index,
